LFS-1232: As a questionnaire designer, I can use text formatting in question and section descriptions Addressed codereview comments
