+++ modules/data-entry/src/main/frontend/src/questionnaireEditor/Question.json
@@ -1,7 +1,7 @@
 [
   {
     "text": "string",
-    "description": "string",
+    "description": "markdown",
     "dataType": {
       "text" : {
         "validationRegexp": "string",
